Extract common functionality of LCSTextDifference into an abstract class.
